refactor: make useGameDataStore only to store game data, for the rest functionalities delegated to dedicated store;
